login

+++ API/FinanceGladiatorProjectApp/Models/tbl_Card.cs
@@ -11,9 +11,9 @@ namespace FinanceGladiatorProjectApp.Models
 {
     using System;
     using System.Collections.Generic;
-    using System.Runtime.Serialization;
+  using System.Runtime.Serialization;
 
-    [DataContract]
+  [DataContract]
     public partial class tbl_Card
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
@@ -21,7 +21,7 @@ namespace FinanceGladiatorProjectApp.Models
         {
             this.tbl_EMI = new HashSet<tbl_EMI>();
         }
-    
+
         [DataMember]
         public int Card_Id { get; set; }
         [DataMember]
@@ -33,7 +33,7 @@ namespace FinanceGladiatorProjectApp.Models
         [DataMember]
         public string Card_Type { get; set; }
         [DataMember]
-        public Nullable<double> Total_credit { get; set; }
+        public Nullable<decimal> Total_credit { get; set; }
         [DataMember]
         public Nullable<int> Card_cost { get; set; }
         [DataMember]
@@ -41,10 +41,12 @@ namespace FinanceGladiatorProjectApp.Models
         [DataMember]
         public Nullable<int> ApprovedBy { get; set; }
         [DataMember]
-        public Nullable<double> credit_used { get; set; }
-    
-        public virtual tbl_Admin tbl_Admin { get; set; }
-        public virtual tbl_Customer tbl_Customer { get; set; }
+        public Nullable<decimal> credit_used { get; set; }
+     
+
+    public virtual tbl_Admin tbl_Admin { get; set; }
+   
+    public virtual tbl_Customer tbl_Customer { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<tbl_EMI> tbl_EMI { get; set; }
     }
